microservices/llm-service/agents/layout_parser.py
import os
import logging
from typing import Dict, List, Any, Optional
from langchain_community.document_loaders import PyMuPDFLoader

logger = logging.getLogger(__name__)

class LayoutParserAgent:
    def __init__(self):
        self.use_doc_ai = False
        # Check for Google Credentials
        if os.getenv("GOOGLE_APPLICATION_CREDENTIALS"):
            try:
                from google.cloud import documentai
                self.use_doc_ai = True
                logger.info("LayoutParserAgent initialized with Google Document AI")
            except ImportError:
                logger.warning("google-cloud-documentai not installed, falling back to PyMuPDF")
        else:
            logger.info("No Google Credentials found, using PyMuPDF fallback")

    def parse_document(self, file_path: str) -> Dict[str, Any]:
        """
        Parses a PDF document and returns a structured representation.
        """
        if self.use_doc_ai:
            return self._parse_with_doc_ai(file_path)
        else:
            return self._parse_with_pymupdf(file_path)

    def _parse_with_doc_ai(self, file_path: str) -> Dict[str, Any]:
        # TODO: Implement actual Document AI logic
        # For now, fallback to PyMuPDF even if credentials exist, until fully implemented
        logger.info("Document AI implementation pending, using PyMuPDF fallback")
        return self._parse_with_pymupdf(file_path)

    def _parse_with_pymupdf(self, file_path: str) -> Dict[str, Any]:
        """
        Fallback parser using PyMuPDF.
        Attempts to identify sections based on simple heuristics (font size, bold text - hard with just PyMuPDFLoader).
        For this MVP, we will treat pages as sections if no better structure is found,
        but we will try to group them.
        """
        logger.info(f"Parsing {file_path} with PyMuPDF")
        loader = PyMuPDFLoader(file_path)
        pages = loader.load()
        
        sections = []
        current_section = {"title": "Introduction", "content": ""}
        
        # Simple heuristic: Group every 5 pages into a "section" to simulate parent blocks
        # In a real implementation, we would use layout analysis to find "Item 1.", "Item 7.", etc.
        pages_per_section = 5
        
        for i, page in enumerate(pages):
            if i > 0 and i % pages_per_section == 0:
                sections.append(current_section)
                current_section = {"title": f"Section {i // pages_per_section + 1}", "content": ""}
            
            current_section["content"] += page.page_content + "\n\n"
            
        sections.append(current_section)
        
        # Extract basic metadata
        filename = os.path.basename(file_path)
        ticker = "UNKNOWN"
        year = "UNKNOWN"
        
        # Try to extract ticker/year from filename if it follows a pattern (e.g., AAPL_2023.pdf)
        parts = filename.replace('.pdf', '').split('_')
        if len(parts) >= 2:
            ticker = parts[0]
            year = parts[1]

        return {
            "filename": filename,
            "ticker": ticker,
            "fiscal_year": year,
            "sections": sections,
            "page_count": len(pages)
        }
