microservices/llm-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
google-cloud-documentai
redis
openai<2.0.0,>=1.10.0
faiss-cpu==1.7.4
pymupdf==1.23.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
requests>=2.25.0
redis>=4.5.0,<5.0.0
python-jose==3.3.0
cryptography==41.0.7
bcrypt==4.0.1
pyjwt==2.8.0
typing-extensions>=4.8.0
anyio<4.0.0,>=3.7.1
pydantic-core==2.10.1
cffi>=1.12
minio==7.1.17
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-logging==0.42b0
jaeger-client==4.8.0